requirements.txt
absl-py==1.4.0
accelerate==0.27.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autofaiss==2.17.0
beartype==0.18.3
bitsandbytes==0.41.1
boto3==1.26.94
botocore==1.29.94
braceexpand==0.1.7
cachetools==5.3.1
charset-normalizer==3.1.0
click==8.1.7
clip-anytorch==2.6.0
cmake==3.25.0
CoCa-pytorch==0.1.0
contourpy==1.1.0
cpm-kernels==1.0.11
cycler==0.11.0
dalle2-pytorch==0.2.4
datasets==2.14.6
deepspeed==0.12.2
dill==0.3.7
docstring-parser==0.15
einops==0.7.0
einops-exts==0.0.4
einx==0.2.0
embedding-reader==1.7.0
faiss-cpu==1.8.0
filelock==3.10.0
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.42.1
frozendict==2.4.2
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.2.0
galore-torch==1.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
hjson==3.1.0
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonlines==4.0.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
libclang==16.0.6
lit==15.0.7
loguru==0.7.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.15
networkx==3.0
ninja==1.11.1.1
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.3
peft==0.9.0
Pillow==9.4.0
protobuf==3.19.6
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-pretrained-bert==0.6.2
pytz==2023.3.post1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
resize-right==0.0.2
rich==13.7.0
rotary-embedding-torch==0.5.3
rouge-chinese==1.0.3
rsa==4.9
s3transfer==0.6.0
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.99
shtab==1.6.5
similarities==1.1.7
six==1.16.0
sympy==1.11.1
termcolor==2.3.0
text2vec==1.2.9
threadpoolctl==3.1.0
tokenizers==0.15.1
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tqdm==4.65.0
transformers==4.38.2
transformers-stream-generator==0.0.4
triton==2.0.0
trl==0.7.6
typing_extensions==4.8.0
tyro==0.7.1
tzdata==2023.3
urllib3==1.26.15
vector_quantize_pytorch==1.14.7
wcwidth==0.2.13
webdataset==0.2.86
Werkzeug==2.3.7
wrapt==1.15.0
x-clip==0.14.4
xxhash==3.4.1
yarl==1.9.2
youtokentome==1.0.6
zipp==3.16.2
